+++ source/MultiGridTable/MultiGridTable.example.js
@@ -1,6 +1,5 @@
 /** @flow */
 import * as React from 'react';
-import clsx from 'clsx';
 import {
   ContentBox,
   ContentBoxHeader,
@@ -8,9 +7,9 @@ import {
 } from '../demo/ContentBox';
 import {LabeledInput, InputRow} from '../demo/LabeledInput';
 import MultiGridTable from './MultiGridTable';
-import styles from './MultiGridTable.example.css';
 import multiGridTableStyles from '../styles.css';
 import {getURL} from '../demo/utils';
+import {Container, FormLabel} from '../demo/Form';
 
 const sample = [
   ['Frozen yoghurt', 159, 6.0, 24, 4.0],
@@ -21,7 +20,7 @@ const sample = [
 ];
 
 function createData(id, dessert, calories, fat, carbs, protein) {
-  return {id, dessert, calories, fat, carbs, protein};
+  return {id: `id-${id}`, dessert, calories, fat, carbs, protein};
 }
 
 const getRows = count => {
@@ -35,51 +34,62 @@ const getRows = count => {
   return rows;
 };
 
-const getColumns = count => {
-  const columns = [
-    {
-      width: 50,
-      label: 'Id',
-      dataKey: 'id',
-      sort: true,
-      align: 'center', // left-right-center default left
-    },
-    {
-      width: 300,
-      label: 'Dessert',
-      dataKey: 'dessert',
-      sort: true,
-      align: 'left',
-    },
-    {
-      width: 120,
-      label: 'Calories\u00A0(g)',
-      dataKey: 'calories',
-      sort: true,
-      align: 'right',
-    },
-    {
-      width: 120,
-      label: 'Fat\u00A0(g)',
-      dataKey: 'fat',
-      sort: true,
-      align: 'right',
-    },
-    {
-      width: 120,
-      label: 'Carbs\u00A0(g)',
-      dataKey: 'carbs',
-      sort: true,
-      align: 'right',
-    },
-    {
-      width: 120,
-      label: 'Protein\u00A0(g)',
-      dataKey: 'protein',
-      sort: false,
-      align: 'right',
-    },
-  ];
+const initColumns = [
+  {
+    width: 50,
+    label: '#',
+    dataKey: 'index',
+    align: 'center', // left-right-center default left
+    render: (value, rowData, rowIndex) => rowIndex + 1,
+  },
+  {
+    width: 50,
+    label: 'Id',
+    dataKey: 'id',
+    sort: true,
+    align: 'left',
+  },
+  {
+    width: 300,
+    label: 'Dessert',
+    dataKey: 'dessert',
+    sort: true,
+    align: 'left',
+  },
+  {
+    width: 120,
+    label: 'Calories\u00A0(g)',
+    dataKey: 'calories',
+    sort: true,
+    align: 'right',
+  },
+  {
+    width: 120,
+    label: 'Fat\u00A0(g)',
+    dataKey: 'fat',
+    sort: true,
+    align: 'right',
+  },
+  {
+    width: 120,
+    label: 'Carbs\u00A0(g)',
+    dataKey: 'carbs',
+    sort: true,
+    align: 'right',
+  },
+  {
+    width: 120,
+    label: 'Protein\u00A0(g)',
+    dataKey: 'protein',
+    sort: false,
+    align: 'right',
+  },
+];
+
+const getColumns = (count = 0) => {
+  const columns = [];
+
+  initColumns.forEach(item => columns.push(item));
 
   for (let i = 0; i < count; i += 1) {
     columns.push({
@@ -94,25 +104,6 @@ const getColumns = count => {
   return columns;
 };
 
-const Container = ({style, children, inline}) => {
-  const className = {[styles.containerInline]: inline};
-  return (
-    <div style={style} className={clsx(styles.container, className)}>
-      {children}
-    </div>
-  );
-};
-
-const FormLabel = ({style, label, children, inline}) => {
-  const className = {[styles.formInline]: inline};
-  return (
-    <div className={clsx(styles.form, className)} style={style}>
-      <div className={styles.label}>{label}</div>
-      <div className={styles.content}>{children}</div>
-    </div>
-  );
-};
-
 export default class MultiGridTableExample extends React.PureComponent {
   constructor(props, context) {
     super(props, context);
@@ -122,17 +113,14 @@ export default class MultiGridTableExample extends React.PureComponent {
       fixedRowCount: 0,
       scrollToColumn: 0,
       scrollToRow: 0,
-      value: {
-        value: '',
-        rowData: {},
-        index: '',
-        dataKey: '',
-      },
+      value: {},
       rowCount: 100,
       columnCount: 0,
       rows: getRows(100),
-      columns: getColumns(0),
+      columns: getColumns(),
       multiple: false,
+      sortBy: {},
+      scroll: {},
     };
 
     this._onScrollToColumnChange = this._createEventHandler('scrollToColumn');
@@ -144,11 +132,9 @@ export default class MultiGridTableExample extends React.PureComponent {
   _createEventHandler(property, func) {
     return event => {
       let value = parseInt(event.target.value, 10) || 0;
-
       if (func) {
         value = func(value);
       }
-
       this.setState({
         [property]: value,
       });
@@ -157,7 +143,6 @@ export default class MultiGridTableExample extends React.PureComponent {
 
   _createLabeledInput(property, eventHandler) {
     const value = this.state[property];
-
     return (
       <LabeledInput
         label={property}
@@ -168,16 +153,20 @@ export default class MultiGridTableExample extends React.PureComponent {
     );
   }
 
-  _onRowClick = (value, rowData, rowIndex, columnKey) => {
-    this.setState({
-      value: {
-        value,
-        rowData,
-        index: rowIndex,
-        dataKey: columnKey,
-      },
-    });
-  };
+  _createLabeledCheckBox(property, eventHandler) {
+    const value = this.state[property];
+    return (
+      <LabeledInput
+        className=""
+        label={property}
+        name={property}
+        onChange={eventHandler}
+        value={value}
+        type="checkbox"
+        checked={value}
+      />
+    );
+  }
 
   _getRows = rowCount => {
     const rows = getRows(rowCount);
@@ -205,6 +194,8 @@ export default class MultiGridTableExample extends React.PureComponent {
       rowCount,
       columnCount,
       multiple,
+      sortBy,
+      scroll,
     } = this.state;
     let {rows, columns} = this.state;
 
@@ -212,17 +203,9 @@ export default class MultiGridTableExample extends React.PureComponent {
       rows = this._getRows(rowCount);
     }
 
-    if (columns.length !== columnCount + 6) {
+    if (columns.length !== columnCount + initColumns.length) {
       columns = this._getColumns(columnCount);
     }
-
-    const checkFixedColumnCount = val => {
-      return val > columns.length - 1 ? columns.length - 1 : val;
-    };
-
-    const checkFixedRowCount = val => {
-      return val > rows.length - 1 ? rows.length - 1 : val;
-    };
 
     return (
       <ContentBox>
@@ -240,11 +223,15 @@ export default class MultiGridTableExample extends React.PureComponent {
         <InputRow>
           {this._createLabeledInput(
             'fixedColumnCount',
-            this._createEventHandler('fixedColumnCount', checkFixedColumnCount),
+            this._createEventHandler('fixedColumnCount', val => {
+              return val > columns.length - 1 ? columns.length - 1 : val;
+            }),
           )}
           {this._createLabeledInput(
             'fixedRowCount',
-            this._createEventHandler('fixedRowCount', checkFixedRowCount),
+            this._createEventHandler('fixedRowCount', val => {
+              return val > rows.length - 1 ? rows.length - 1 : val;
+            }),
           )}
           {this._createLabeledInput(
             'scrollToColumn',
@@ -256,45 +243,85 @@ export default class MultiGridTableExample extends React.PureComponent {
             'columnCount',
             this._onScrollColumnCountChange,
           )}
+          {this._createLabeledCheckBox('multiple', event => {
+            this.setState({multiple: event.target.checked});
+          })}
         </InputRow>
 
         <Container style={{marginTop: 5}}>
-          <FormLabel label="value:" inline>
-            <code>{JSON.stringify(value.value)}</code>
-          </FormLabel>
+          <Container inline>
+            <FormLabel label="value:" inline>
+              {value.value}
+            </FormLabel>
+            <FormLabel label="index:" inline>
+              {value.index}
+            </FormLabel>
+            <FormLabel label="dataKey:" inline>
+              {value.dataKey}
+            </FormLabel>
+          </Container>
           <FormLabel label="rowData:" inline>
-            <code>{JSON.stringify(value.rowData)}</code>
+            <code>{JSON.stringify(value.rowData, undefined, 2)}</code>
           </FormLabel>
-          <FormLabel label="index:" inline>
-            {value.index}
+          <FormLabel label="sortBy:" inline>
+            <code>{JSON.stringify(sortBy, undefined, 2)}</code>
           </FormLabel>
-          <FormLabel label="rowData:" inline>
-            {value.dataKey}
-          </FormLabel>
-          <FormLabel label="multiple mode:" inline>
-            <input
-              aria-label="multiple"
-              type="checkbox"
-              checked={multiple}
-              onChange={event =>
-                this.setState({multiple: event.target.checked})
-              }
-            />
+          <FormLabel label="scroll:" inline>
+            <code>{JSON.stringify(scroll, undefined, 2)}</code>
           </FormLabel>
         </Container>
 
         <div style={{width: '100%', height: 360}}>
           <MultiGridTable
-            rows={rows}
-            columns={columns}
             fixedColumnCount={fixedColumnCount}
             fixedRowCount={fixedRowCount}
             scrollToColumn={scrollToColumn}
             scrollToRow={scrollToRow}
-            onRowClick={this._onRowClick}
+            classes={multiGridTableStyles}
+            rows={rows}
+            columns={columns}
             value={value.value}
             multiple={multiple}
-            classes={multiGridTableStyles}
+            classNameCell={(id, rowData, rowIndex) => {
+              return `row-cell-${rowIndex}`;
+            }}
+            onRowClick={(value, rowData, rowIndex, columnKey, event) => {
+              this.setState({
+                value: {
+                  value,
+                  rowData,
+                  index: rowIndex,
+                  dataKey: columnKey,
+                },
+              });
+            }}
+            onHeaderRowClick={({sortBy, sortDirection}) => {
+              this.setState({
+                sortBy: {
+                  sortBy,
+                  sortDirection,
+                },
+              });
+            }}
+            onScroll={({
+              clientHeight,
+              clientWidth,
+              scrollHeight,
+              scrollLeft,
+              scrollTop,
+              scrollWidth,
+            }) => {
+              this.setState({
+                scroll: {
+                  clientHeight,
+                  clientWidth,
+                  scrollHeight,
+                  scrollLeft,
+                  scrollTop,
+                  scrollWidth,
+                },
+              });
+            }}
           />
         </div>
       </ContentBox>